Exceptionless telemeters are added

+++ TIKSN.Core/Analytics/Telemetry/CompositeMetricTelemeter.cs
@@ -14,7 +14,7 @@ namespace TIKSN.Analytics.Telemetry
 			this.metricTelemeters = metricTelemeters;
 		}
 
-		public async Task TrackMetric(string metricName, double metricValue)
+		public async Task TrackMetric(string metricName, decimal metricValue)
 		{
 			if (commonConfiguration.GetConfiguration().IsMetricTrackingEnabled)
 			{
